chore(Store): Service 인터페이스 추가, Service 구현체 이름 수정
